Move get{Bind,Tmpfs,Secret,SSH}Mount to run_common.go
